// app/api/list-recipe/shop-unit/schema.ts
import { z } from "zod";

export const PatchShopUnitSchema = z.object({
    ingredient_id: z.string().min(1),
    connection_id: z.string().min(1),
    _id: z.string().min(1),
    unit_id: z.string().min(1),
    shop_unit: z.boolean(),
}).strict();

export type PatchShopUnitDTO = z.infer<typeof PatchShopUnitSchema>;